fix all bugs and finished experiments

+++ db/log_structured.h
@@ -90,9 +90,11 @@ class LogStructured {
 
   std::vector<std::vector<LogSegment *>> class_segments_{num_class};
 
-  const int class1_sz = 320/2; // MB
-  const int class2_sz = 240/2; // MB
-  std::vector<int> class_sz = {0, class1_sz, class2_sz};
+  const uint64_t class1_sz = 320/2; // MB
+  const uint64_t class2_sz = 240/2; // MB
+  // const uint64_t class1_sz = (1 << 10) * 4; // MB
+  // const uint64_t class2_sz = (1 << 9) * 4; // MB
+  std::vector<uint64_t> class_sz = {0, class1_sz, class2_sz};
 
   std::atomic<int> alloc_counter_{0};
   const int num_limit_free_segments_;
